backend/requirements.txt
Flask
flask_cors #eliminar para prod
Flask-JWT-Extended
Flask-SQLAlchemy
mysql-connector-python
gunicorn
timedelta